requirements.txt
pydantic
pytesseract
Flask
Pillow
pdf2image
requests
openai
iso4217
